% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_add_tidal_data.R
\name{atl_add_tidal_data}
\alias{atl_add_tidal_data}
\title{Add tidal data to tracking data}
\usage{
atl_add_tidal_data(
  data,
  tide_data,
  tide_data_highres,
  waterdata_resolution = "10 min",
  waterdata_interpolation = NULL,
  offset = 0
)
}
\arguments{
\item{data}{A dataframe with the tracking data with the timestamp column
'datetime' in UTC.}

\item{tide_data}{Data on the timing (in UTC) of low and high tides.}

\item{tide_data_highres}{Data on the timing (in UTC) of the water level in
small intervals (e.g. every 10 min) as provided from Rijkswaterstaat.}

\item{waterdata_resolution}{The resolution of the high resolution water level
data. This is used for matching the high resolution tidal data to the
tracking data. Defaults to 10 minutes but can be set differently.}

\item{waterdata_interpolation}{Time interval to which the water level data
will be interpolated (should be smaller than water data resolution e.g.
1 min). If NULL will keep the original water data resolution.}

\item{offset}{The offset in minutes between the location of the tidal gauge
and the tracking area. This value will be added to the timing of the
water data.}
}
\value{
The input data but with three columns added: tideID (a unique number
for the tidal period between two consecutive high tides), tidaltime (time
since high tide in minutes), time2lowtide (time to low tide in minutes),
and water level with reference to NAP (cm).
}
\description{
Adds a unique tide identifier, water level, time from high tide and time to
low tide for tracking data (both in minutes).
}
\examples{
# packages
library(tools4watlas)

# load example data
data <- data_example

# delete existing tide data columns to show how they are added
data[, c("tideID", "tidaltime", "time2lowtide", "waterlevel") := NULL]

# sub path to tide data
tidal_pattern_fp <- system.file(
  "extdata", "example-tidalPattern-west_terschelling-UTC.csv",
  package = "tools4watlas"
)
measured_water_height_fp <- system.file(
  "extdata", "example-gemeten_waterhoogte-west_terschelling-clean-UTC.csv",
  package = "tools4watlas"
)

# load tide data
tidal_pattern <- fread(tidal_pattern_fp)
measured_water_height <- fread(measured_water_height_fp)

# add tide data to movement data
# The offset of 30 minutes is set to match Griend.
data <- atl_add_tidal_data(
  data = data,
  tide_data = tidal_pattern,
  tide_data_highres = measured_water_height,
  waterdata_resolution = "10 min",
  waterdata_interpolation = "1 min",
  offset = 30
)

# show first 5 rows (subset of columns to show additional ones)
head(data[, .(tag, datetime, tideID, tidaltime, time2lowtide, waterlevel)])
}
\author{
Pratik Gupte & Allert Bijleveld & Johannes Krietsch
}
