% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_bbox.R
\name{atl_bbox}
\alias{atl_bbox}
\title{Create a bounding box with specified aspect ratio and buffer}
\usage{
atl_bbox(data, x = "x", y = "y", asp = "16:9", buffer = 0)
}
\arguments{
\item{data}{An \code{sf} or \code{sfc} object for which the bounding box is
calculated or a data.table with x- and y- coordinates.}

\item{x}{A character string representing the name of the column containing
x-coordinates. Defaults to "x".}

\item{y}{A character string representing the name of the column containing
y-coordinates. Defaults to "y".}

\item{asp}{A character string specifying the desired aspect ratio in the
format \code{"width:height"}. Default is \code{"16:9"}, if \code{NULL} returns simple
bounding box without modifying aspect ratio.}

\item{buffer}{A numeric value (in meters) specifying the buffer distance to
be applied to the bounding box. Positive values expand the bounding box,
while negative values shrink it. Default is \code{0}.}
}
\value{
A bounding box (\code{bbox}), represented as a named vector with
\code{xmin}, \code{ymin}, \code{xmax}, and \code{ymax} values.
}
\description{
This function generates a bounding box for a given geometry with a
specified aspect ratio. Additionally, it allows applying a buffer to
expand or contract the bounding box.
}
\examples{
# packages
library(ggplot2)
library(sf)

# load example data
data <- data_example

# bounding box based on data
bbox <- atl_bbox(data, buffer = 1000)

# bounding box based on specified coordinates in EPSG:4326
bbox <- data.table(x = c(5.107, 5.330), y = c(53.303, 53.230)) |>
  st_as_sf(coords = c("x", "y"), crs = 4326) |>
  st_transform(crs = 32631) |>
  atl_bbox(buffer = 1000)

# bounding box based on polygon
bbox <- atl_bbox(grienderwaard, buffer = 1000)

# create basemap with bounding box
bm <- atl_create_bm(bbox)

# plot bm with bounding box when bm coordinates are overridden by geom_sf
bm +
  geom_sf(data = grienderwaard, fill = "transparent", color = "firebrick") +
  # set extend again (overwritten by geom_sf)
  coord_sf(
    xlim = c(bbox["xmin"], bbox["xmax"]),
    ylim = c(bbox["ymin"], bbox["ymax"]), expand = FALSE
  )
}
\author{
Johannes Krietsch
}
