% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_file_path.R
\name{atl_file_path}
\alias{atl_file_path}
\title{Get the file path for WATLAS or GIS data based on the user's name.}
\usage{
atl_file_path(data_type = c("watlas_teams", "rasters", "shapefiles"))
}
\arguments{
\item{data_type}{A character string indicating the type of data. Options are:
\itemize{
\item \code{"watlas_teams"}: Path to “WATLAS” SharePoint folder: Documents/data/
\item \code{"rasters"}: Path to “Birds, fish ’n chips” SharePoint folder:
Documents/data/GIS/rasters/
\item \code{"shapefiles"}: Path to “Birds, fish ’n chips” SharePoint folder:
Documents/data/GIS/shapefiles/
}}
}
\value{
A character string representing the full file path to the selected
data type for the current user.
}
\description{
This function returns a predefined file path based on the user's system
username and the selected data type.
Users must be predefined in the function.
}
\details{
For local users with access to the NIOZ network, I would recommend to add a
shortcut to your OneDrive of the “WATLAS” SharePoint folder: Documents/data/.
The path in the function then refers to this copy on your OneDrive. Files in
this folder (e.g. tags_watlas_all.xlsx and tide data) are often updated and
changes are then automatically updated on your computer.

The second source of data is the “Birds, fish ’n chips” SharePoint folder:
Documents/data/GIS/ which has the subfolders rasters and shapefiles. One can
also create a OneDrive shortcut for this folder, however it contains many
large files and files usually don't change. I would therefore recommend to
only copy necessary data in a local folder on your computer and then link
the path to this folder.

To build all articles for the package website, you need to specify your user
paths here.
}
