% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_full_tag_id.R
\name{atl_full_tag_id}
\alias{atl_full_tag_id}
\title{Create full tag name}
\usage{
atl_full_tag_id(tag, short = FALSE)
}
\arguments{
\item{tag}{tag ID, either numeric or character}

\item{short}{TRUE or FALSE for short or long tag ID}
}
\value{
Full or short tag ID as character
}
\description{
Create full tag name
}
\examples{

tag = 123

atl_full_tag_id(tag)
atl_full_tag_id(tag, short = TRUE)

}
