% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxiliary.r
\name{atl_get_spatial_bounds}
\alias{atl_get_spatial_bounds}
\title{Get spatial bounds for dataframe X,Y coordinates.}
\usage{
atl_get_spatial_bounds(data)
}
\arguments{
\item{data}{A dataframe with the tracking data.}
}
\value{
Provides a matrix with the range in X and Y coordinates.
}
\description{
Obtains the extent of the (localization) data within a dataframe that contains coordinates X and Y.
}
\author{
Allert Bijleveld
}
