% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxiliary.R
\name{atl_make_boundingbox_for_osm}
\alias{atl_make_boundingbox_for_osm}
\title{Create bounding box in LatLong for downloading a map from OpenStreetMap.}
\usage{
atl_make_boundingbox_for_osm(
  bbox,
  buffer = 1000,
  from_crs = sp::CRS("+init=epsg:32631")
)
}
\arguments{
\item{bbox}{A matrix with the ranges in X and Y coordinates. The first
column contains the range in X, and the second column the range in Y.
The first row contains the minimum and the second row the maximum values.}

\item{buffer}{The buffer (in meters) for extending the \code{bbox}.}

\item{from_crs}{The CRS() of the bbox.}
}
\value{
Provides a data.frame that can be used with \code{openmap()} where
the first row provides the upperleft corner and the second row the
lowerright corner of the extent, and the first column refers to the
Y-coordinates and the second column to the X-coordinates.
}
\description{
Transforms a bounding box obtained with \code{atl_get_spatial_bounds} to a
bounding box in LongLat necesarry for plotting with
\code{OpenStreetMap::openmap()} with the function \code{atl_plot_tag}.
}
\author{
Allert Bijleveld
}
