% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxiliary.r
\name{atl_make_spatialdataframe}
\alias{atl_make_spatialdataframe}
\title{Make X,Y data spatial}
\usage{
atl_make_spatialdataframe(data, crs = sp::CRS("+init=epsg:32631"))
}
\arguments{
\item{data}{A dataframe with the tracking data.}

\item{crs}{The coordinate reference system (specified with \code{CRS()}) of
the X,Y coordinates. The deafult is UTM 31N: \code{CRS("+init=epsg:32631")}.}
}
\value{
The output as a Spatialpointsdataframe.
}
\description{
A function that will use the library sp to convert to spatialdataframe.
}
\author{
Allert Bijleveld
}
