% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_patch_summary.R
\name{atl_patch_summary}
\alias{atl_patch_summary}
\title{Summary of patch data}
\usage{
atl_patch_summary(
  data,
  buffer = 15,
  summary_variables = c(),
  summary_functions = c()
)
}
\arguments{
\item{buffer}{A numeric value (in meters) specifying the buffer distance for
the bounding box. Default is 15 m, but could for example be
\code{lim_spat_indep} of the residency patch calculation.}

\item{summary_variables}{Optional variables for which patch-wise summary
values are required. To be passed as a character vector.}

\item{summary_functions}{The functions with which to summarise the summary
variables; must return only a single value, such as median, mean etc. To be
passed as a character vector.}
}
\value{
A data.table that has the added column
\code{patch}, \code{patchdata}, and \code{polygons}, indicating the patch
identity, the localization data used to construct the patch, and the polygons
of residence patches based on the \code{lim_spat_indep}. In addition, there
are columns with patch summaries: nfixes, dist_in_patch, dist_bw_patch and
statistics based on the \code{summary_variables} and \code{summary_functions}
provided.
}
\description{
The function \code{atl_patch_summary} can be used to extract patch-specific
summary data such as the median coordinates, the patch duration, the distance
travelled within the patch, the displacement within the patch, and the patch
area. Position covariates such as speed may also be
summarised patch-wise by passing covariate names and  summary functions as
character vectors to the \code{summary_variables} and
\code{summary_functions} arguments, respectively.
}
