% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxiliary.r
\name{atl_plot_add_track}
\alias{atl_plot_add_track}
\title{Add tracks to plot from list}
\usage{
atl_plot_add_track(
  data,
  Pch = 19,
  Cex = 0.25,
  Lwd = 1,
  col,
  Type = "o",
  endpoint = FALSE
)
}
\arguments{
\item{data}{The spatial data frame.}

\item{Pch}{The type of point to plot a localization}

\item{Cex}{The size of the point to plot a localization}

\item{Lwd}{The width of the line to connect localizations}

\item{col}{The colour of plotted localizations}

\item{Type}{The type of graph to make. For instance, "b" is both points
and lines and "o" is simlar but places points on top of line (no gaps)}

\item{endpoint}{Whether to plot the last localization of an individual
in magenta}
}
\description{
A function that is used for plotting multiple individuals on a map from a
list of spatial data.
}
\author{
Allert Bijleveld
}
