% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxiliary.R
\name{atl_plot_map_osm}
\alias{atl_plot_map_osm}
\title{Plot a map downloaded with OpenStreetMap}
\usage{
atl_plot_map_osm(map, ppi = 96)
}
\arguments{
\item{map}{The map loaded with \code{OpenStreetMap::openmap()}.}

\item{ppi}{The pixels per inch, which is used to calculate the dimensions of
the plotting region from \code{mapID}. Deafults to 96.}
}
\value{
Returns an OSM background plot for adding tracks.
}
\description{
A function that is used in e.g. plotting multiple individuals.
}
\author{
Allert Bijleveld
}
