% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxiliary.r
\name{atl_plot_rpatches}
\alias{atl_plot_rpatches}
\title{Add residence patches to a plot}
\usage{
atl_plot_rpatches(
  data,
  Pch = 21,
  Cex = 0.25,
  Lwd = 1,
  Col = 1,
  Bg = NULL,
  Lines = TRUE
)
}
\arguments{
\item{data}{Either sfc_Polygon or a dataframe with the tracking data}

\item{Pch}{Corresponding graphical argument passed on to the base plot
function}

\item{Cex}{Corresponding graphical argument passed on to the base plot
function}

\item{Lwd}{Corresponding graphical argument passed on to the base plot
function}

\item{Col}{Corresponding graphical argument passed on to the base plot
function}

\item{Bg}{Corresponding graphical argument passed on to the base plot
function}

\item{Lines}{Corresponding graphical argument passed on to the base plot
function}
}
\value{
Nothing but an addition to the current plotting device.
}
\description{
Adds residence pattch data in UTM 31N as points or polygons to a plot.
}
\author{
Allert Bijleveld
}
