% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot_tag.r
\name{atl_plot_tag}
\alias{atl_plot_tag}
\title{Plot track for one individual on a simple background.}
\usage{
atl_plot_tag(
  data,
  tag = NULL,
  fullname = NULL,
  color_by = "time",
  towers = NULL,
  h = 7,
  w = 7 * (16/9),
  buffer = 1,
  Legend = "topleft",
  Scalebar = 5,
  cex_legend = 1
)
}
\arguments{
\item{data}{A dataframe with the tracking data. Can include multiple tags, but one tag is selected for plotting.}

\item{tag}{The four-digit tag number as character to plot. Defaults to plotting the first tag in \code{data}.}

\item{fullname}{If specified the plot will be saved in this path with this name (include extension). Defaults to NULL and plotting in a graphics window.}

\item{color_by}{Either \code{"time"}, \code{"SD"}, or \code{"NBS"}, which are respectively used to colour the localization with the relative time (hours), variance in the localizations as the maximum of VARX and VARY, or the Number of Base Stations (NBS) used to calculate the localization. Defaults to "time".}

\item{towers}{A dataframe with coordinates of receiver stations (named \code{X} and \code{Y}).}

\item{Legend}{Passed to the \code{legend} function and sets the location of the legend in the plot.}

\item{Scalebar}{Length of scalebar in km.}

\item{cex_legend}{The size of the text in the legend.}
}
\value{
Returns nothing but a plot.
}
\description{
A function that plots the localization data of one individual.
}
\author{
Allert Bijleveld
}
