% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_animate_data.R
\name{atl_progress_bar}
\alias{atl_progress_bar}
\title{Display a live progress bar for PNG file generation in a directory}
\usage{
atl_progress_bar(file_path, total = NULL, refresh_rate = 1, quietly = FALSE)
}
\arguments{
\item{file_path}{Path to the directory containing PNG files.}

\item{total}{Optional. Total number of expected PNG files. If NULL, the
function reads from 'total_frames.txt' created by atl_time_steps() in the
same directory.}

\item{refresh_rate}{Numeric value in seconds specifying how often the
progress bar updates.}

\item{quietly}{Logical. If \code{TRUE}, no progress or completion messages are
printed. Default is \code{FALSE}.}
}
\value{
No return value. Prints progress bar to the console.
}
\description{
This function is meant to track the progress of PNG's created in a parallel
loop. It will check the number of PNG files in a specified directory and
make a progress bar in the console. To use the function, open a new R session
and run the function there.
}
