% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_res_patch.r
\name{atl_res_patch}
\alias{atl_res_patch}
\title{Construct residence patches from position data}
\usage{
atl_res_patch(
  data,
  max_speed = 3,
  lim_spat_indep = 75,
  lim_time_indep = 180,
  min_fixes = 2,
  min_duration = 60
)
}
\arguments{
\item{data}{A dataframe of any class that is or extends data.frame of one
individual only. The dataframe must contain at least two spatial coordinates,
\code{x} and \code{y}, and a temporal coordinate, \code{time}.}

\item{max_speed}{A numeric value specifying the maximum speed (m/s) between
two coordinates that would be considered non-transitory}

\item{lim_spat_indep}{A numeric value of distance in metres of the spatial
distance between two patches for them to the considered independent.}

\item{lim_time_indep}{A numeric value of time in minutes of the time
difference between two patches for them to be considered independent.}

\item{min_fixes}{The minimum number of fixes for a group of
spatially-proximate number of points to be considered a preliminary residence
patch.}

\item{min_duration}{The minimum duration (in seconds) for classifying
residence patches.}
}
\value{
A data.table that has the added column
\code{patch} as character indicating the patch ID.
}
\description{
A cleaned movement track of one individual at a time can be classified into
residence patches using the
function \code{atl_res_patch}.
The function expects a specific organisation of the data: there should be
at least the following columns, \code{x}, \code{y}, and \code{time},
corresponding to the coordinates, and the time as \code{POSIXct}.
\code{atl_res_patch} requires only three parameters: (1) the maximum
speed threshold between localizations (called \code{max_speed}), (2) the
distance threshold between proto-patches of positions (called
\code{lim_spat_indep}), and (3) the time interval between proto-patches
(called \code{lim_time_indep}). As the code initially only looks at
proto-patches, at the end it checks if positions within patches are
interrupted by short flights (with a distance larger than
\code{lim_spat_indep} to the last position of the proto-patch before and
first position of the next proto-patch). If there are more than
\code{min_fixes} in this bout, then the patch will be split. If there are
less, we assume this to be single outliers and only assign no patch ID
}
\details{
Derived from \code{atlastools::atl_res_patch()} in the \pkg{atlastools}
package (Gupte et al., 2022), licensed under GPL-3.
}
\examples{
# packages
library(tools4watlas)

# load example data
data <- data_example

# calculate residence patches for one red knot
data <- atl_res_patch(
  data[tag == "3038"],
  max_speed = 3, lim_spat_indep = 75, lim_time_indep = 180,
  min_fixes = 2, min_duration = 60
)

# summary of residence patches
data_summary <- atl_res_patch_summary(data)
data_summary
}
\references{
Gupte, P. R., Beardsworth, C. E., Spiegel, O., Lourie, E., Toledo, S.,
Nathan, R., & Bijleveld, A. I. (2022). A guide to pre-processing
high-throughput animal tracking data. \emph{Journal of Animal Ecology}, 91,
287-307. \doi{10.1111/1365-2656.13610}
}
\author{
Pratik R. Gupte, Christine E. Beardsworth, Allert I. Bijleveld &
Johannes Krietsch
}
