% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make_res_patch_speed.r
\name{atl_res_patch_speed}
\alias{atl_res_patch_speed}
\title{Construct residence patches from position data.}
\usage{
atl_res_patch_speed(
  data,
  max_speed = 3,
  lim_spat_indep = 75,
  lim_time_indep = 180,
  min_fixes = 3,
  min_duration = 120,
  summary_variables = c(),
  summary_functions = c()
)
}
\arguments{
\item{data}{A dataframe of any class that is or extends data.frame of one individual only.
The dataframe must contain at least two spatial coordinates, \code{X} and
\code{Y}, and a temporal coordinate, \code{time}.}

\item{max_speed}{A numeric value specifying the maximum speed (m/s) between two
coordinates that would be considered non-transitory}

\item{lim_spat_indep}{A numeric value of distance in metres of the spatial
distance between two patches for them to the considered independent.}

\item{lim_time_indep}{A numeric value of time in minutes of the time
difference between two patches for them to be considered independent.}

\item{min_fixes}{The minimum number of fixes for a group of
spatially-proximate number of ponts to be considered a preliminary residence
patch.}

\item{min_duration}{The minimum duration (in seconds) for classifying residence patches.}

\item{summary_variables}{Optional variables for which patch-wise summary
values are required. To be passed as a character vector.}

\item{summary_functions}{The functions with which to summarise the summary
variables; must return only a single value, such as median, mean etc. To be
passed as a character vector.}
}
\value{
A data.frame extension object. This dataframe has the added column
\code{patch}, \code{patchdata}, and \code{polygons}, indicating the patch identity, the
localization data used to construct the patch, and the polygons of residence patches based on the \code{lim_spat_indep}. In addition, there are columns with patch summaries: nfixes, dist_in_patch, dist_bw_patch and statistics based on the \code{summary_variables} and \code{summary_functions} provided.
summary variables.
}
\description{
A cleaned movement track of one individual at a time can be classified into residence patches using the
function \code{atl_res_patch_speed}.
The function expects a specific organisation of the data: there should be
at least the following columns, \code{X}, \code{Y}, and \code{time}, corresponding to the coordinates, and the time as \code{POSIXct}.
\code{atl_res_patch_speed} requires only three parameters: (1) the maximum speed threshold between localizations (called \code{max_speed}), (2) the distance
threshold between clusters of positions (called \code{lim_spat_indep}),
and (3) the time interval between clusters (called \code{lim_time_indep}).
Clusters formed of fewer than a minimum number of positions can be excluded.
The exclusion of clusters with few positions can help in removing bias due to
short stops, but if such short stops are also of interest, they can be
included by reducing the \code{min_fixes} argument.
Position covariates such as speed may also be summarised patch-wise by
passing covariate names and  summary functions as character vectors to the
\code{summary_variables} and \code{summary_functions} arguments, respectively.
}
\examples{
require(data.table)
require(ggplot2)

# Load data
d = atl_get_data_csv() |> data.table()
setnames(d, c("X", "Y"), c("x", "y"))

# Calculate residency patches
rp = atl_res_patch_speed(
  d, max_speed = 3, lim_spat_indep = 75, lim_time_indep = 180, min_fixes = 3, 
  min_duration = 120,
  summary_functions = c("mean", "median", "sd", "min", "max", "first", "last"))

# Extract data from all residency patches
dr = rp[, rbindlist(lapply(patchdata, function(x) cbind(x))), by = patch]

# Merge with data
d = merge(d, dr[, .(TAG, posID, patch)], by = c("TAG", "posID"), all.x = TRUE)

# Plot data
ggplot(d) +
  geom_path(aes(x, y), alpha = 0.1) +
  geom_point(aes(x, y, color = as.character(patch)), show.legend = FALSE) +
  theme_bw()
}
\author{
Pratik R. Gupte, Christine E. Beardsworth & Allert I. Bijleveld
}
