% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_animate_data.R
\name{atl_size_along}
\alias{atl_size_along}
\title{Creates different size values along a vector}
\usage{
atl_size_along(x, head = 20, to = c(0.1, 2.5))
}
\arguments{
\item{x}{Vector along which alpha is created}

\item{head}{Numeric parameter influencing the length of the head}

\item{to}{Numeric vector including the minimum and maximum size}
}
\value{
Numeric vector with different size values
}
\description{
Copied from https://github.com/ornitho-logics/windR
}
\examples{
library(ggplot2)
d <- data.frame(
  x = 1:100, y = 1:100,
  s = atl_size_along(1:100, head = 70, to = c(0.1, 5))
)
bm <- ggplot(d, aes(x, y))
bm + geom_path(linewidth = 1)
bm + geom_path(linewidth = d$s, lineend = "round")
}
\author{
Mihai Valcu & Johannes Krietsch
}
