% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_summary.R
\name{atl_summary}
\alias{atl_summary}
\title{Summary of localization data}
\usage{
atl_summary(data, id_columns = c("tag"))
}
\arguments{
\item{data}{A data.table containing localization data with columns for
ID, x, y coordinates, time, and datetime.}

\item{id_columns}{A character vector specifying the column(s) to group by.
Defaults to "tag".}
}
\value{
A data.table with summary statistics for each ID group, including
the total number of positions, first and last data, days of data and time
gaps between localizations, as well as  data coverage.
}
\description{
This function generates a summary of localization data by calculating the
total number of positions, first and last data, days of data and time gaps
between localizations, as well as  data coverage. It returns a summary for
each unique ID specified in \code{id_columns}.
}
\examples{
# packages
library(tools4watlas)

# path to csv with filtered data
data_path <- system.file(
  "extdata", "watlas_data_filtered.csv",
  package = "tools4watlas"
)

# load data
data <- fread(data_path, yaml = TRUE)

# summarize data
summary <- atl_summary(data, id_columns = c("tag"))
summary
}
\author{
Johannes Krietsch
}
