% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_animate_data.R
\name{atl_time_steps}
\alias{atl_time_steps}
\title{Generate time steps and file names for an animation of movements}
\usage{
atl_time_steps(
  datetime_vector,
  time_interval = "10 min",
  output_path,
  create_path = FALSE,
  fps = 24
)
}
\arguments{
\item{datetime_vector}{A vector of datetime values (POSIXct or similar).
Can be a min and max or simple a full vector from the data}

\item{time_interval}{A character string specifying the time interval
(e.g., "30 sec", "10 min", "1 hour").}

\item{output_path}{A character string specifying the directory of the folder
where the files will be saved.}

\item{create_path}{A logical value. If TRUE, the function creates the
directory if it does not exist.}

\item{fps}{A numeric value specifying the frames per second (fps). Only used
to calculate the duration of the final animation. The frame rate needs to be
specified in ffmpeg.}
}
\value{
A data.table with two columns:
\itemize{
\item \code{datetime}: The generated time steps.
\item \code{path}: Corresponding file paths for each time step.
}
}
\description{
This function creates a sequence of time steps based on a given datetime
vector and time interval. It also generates corresponding file names in a
provided folder path for each time step.
}
\examples{
library(tools4watlas)

# load example data
data <- data_example

# create time steps
ts <- atl_time_steps(
  datetime_vector = data$datetime,
  time_interval = "10 min",
  output_path = tempdir(),
  create_path = FALSE
)
ts
}
\author{
Johannes Krietsch
}
