% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_transform_dt.R
\name{atl_transform_dt}
\alias{atl_transform_dt}
\title{Transform coordinates in a data.table and appends new EPSG-Suffixed columns}
\usage{
atl_transform_dt(
  data,
  x = "x",
  y = "y",
  from = sf::st_crs(32631),
  to = sf::st_crs(4326)
)
}
\arguments{
\item{data}{A \code{data.table} containing coordinate columns.}

\item{x}{A character string specifying the column with x-coordinates.
Defaults to \code{"x"}.}

\item{y}{A character string specifying the column with y-coordinates.
Defaults to \code{"y"}.}

\item{from}{An \code{sf::st_crs} object representing the source coordinate
reference system.
Defaults to EPSG:32631.}

\item{to}{An \code{sf::st_crs} object representing the target coordinate
reference system.
Defaults to EPSG:4326.}
}
\value{
A \code{data.table} identical to the input but with two new columns containing the
transformed coordinates.
}
\description{
Transforms coordinate columns in a \code{data.table} from one CRS to another using
\strong{sf}, and appends the transformed coordinates as new columns.
The new columns are automatically named using the original column names
suffixed with the EPSG code of the target CRS (e.g., \code{x_4326}, \code{y_4326}).
Original coordinates are preserved.
}
\examples{
# packages
library(tools4watlas)

# example with bbox from data and movement data
data <- data_example

# add transformed coordinates in projection of the base map (EPSG:4326)
data <- atl_transform_dt(data)

# check data
data[, .(tag, datetime, x, y, x_4326, y_4326)]
}
\author{
Johannes Krietsch
}
