% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lakes_sf}
\alias{lakes_sf}
\title{OSM data lakes around Griend}
\format{
sf object used for background layer in the function atl_create_bm()
}
\source{
\url{https://www.openstreetmap.org}
}
\usage{
lakes_sf
}
\description{
OSM data lakes around Griend
}
\keyword{datasets}
