% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land}
\alias{land}
\title{Land polygon around the Dutch Wadden Sea}
\format{
sf object used for background layer in the function atl_create_bm()
}
\source{
\url{https://www.openstreetmap.org}
}
\usage{
land
}
\description{
Land polygon around the Dutch Wadden Sea
}
\keyword{datasets}
