% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{land_sf}
\alias{land_sf}
\title{OSM data land around Griend}
\format{
sf object used for background layer in the function atl_create_bm()
}
\source{
\url{https://www.openstreetmap.org}
}
\usage{
land_sf
}
\description{
OSM data land around Griend
}
\keyword{datasets}
