% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mudflats_sf}
\alias{mudflats_sf}
\title{OSM data mudflats around Griend}
\format{
sf object used for background layer in the function atl_create_bm()
}
\source{
\url{https://www.openstreetmap.org}
}
\usage{
mudflats_sf
}
\description{
OSM data mudflats around Griend
}
\keyword{datasets}
