% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxiliary.r
\name{plot_map_osm}
\alias{plot_map_osm}
\title{Plot a map downloaded with OpenStreetMap.}
\usage{
plot_map_osm(map, ppi = 96)
}
\arguments{
\item{map}{The map loaded with \code{OpenStreetMap::openmap()}.}
}
\value{
Returns an OSM background plot for adding tracks.
}
\description{
A function that is used in e.g. plotting multiple individuals.
}
\author{
Allert Bijleveld
}
