% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rivers_sf}
\alias{rivers_sf}
\title{OSM data rivers around Griend}
\format{
sf object used for background layer in the function atl_create_bm()
}
\source{
\url{https://www.openstreetmap.org}
}
\usage{
rivers_sf
}
\description{
OSM data rivers around Griend
}
\keyword{datasets}
