% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_add_raster_data.R
\name{atl_add_raster_data}
\alias{atl_add_raster_data}
\title{Add raster data to tracking data}
\usage{
atl_add_raster_data(
  data = NULL,
  x = "x",
  y = "y",
  projection = sf::st_crs(32631),
  raster_data,
  var_name = NULL,
  new_name = NULL,
  change_unit = 1
)
}
\arguments{
\item{data}{A \code{data.table} containing the data to which raster values will
be added. If not a \code{data.table}, it will be coerced to one.}

\item{x}{Character string specifying the column name for x-coordinates.
Defaults to \code{"x"}.}

\item{y}{Character string specifying the column name for y-coordinates.
Defaults to \code{"y"}.}

\item{projection}{A coordinate reference system (CRS) for the spatial data
in the input. Defaults to EPSG:32631.}

\item{raster_data}{A \code{SpatRaster} object from which values will be extracted.}

\item{var_name}{Character string specifying the raster variable to extract.
Defaults to the first layer if \code{NULL}.}

\item{new_name}{Character string specifying the name of the new column in the
output. If \code{NULL}, uses \code{var_name}.}

\item{change_unit}{Numeric value by which to multiply extracted raster values
#'   before adding them to the data. Defaults to \code{1}.}
}
\value{
A \code{data.table} with the extracted raster data added as a new column.
}
\description{
This function extracts raster data (for example bathymetry data) at specified
coordinates and adds it as a column to the input data.table.
}
