% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_as_sf.R
\name{atl_as_sf}
\alias{atl_as_sf}
\title{Convert a data.frame or data.table to an simple feature (sf) object}
\usage{
atl_as_sf(
  data,
  tag = "tag",
  x = "x",
  y = "y",
  projection = sf::st_crs(32631),
  additional_cols = NULL,
  option = "points",
  buffer
)
}
\arguments{
\item{data}{A \code{data.table} or an object convertible to a \code{data.table}.
The input data containing the coordinates and optional attributes.}

\item{tag}{A character string representing the  name of the column containing
the tag ID.}

\item{x}{A character string representing the name of the column containing
x-coordinates. Defaults to "x".}

\item{y}{A character string representing the name of the column containing
y-coordinates. Defaults to "y".}

\item{projection}{An object of class \code{crs} representing the coordinate
reference system (CRS) to assign to the resulting sf object. Defaults to
EPSG:32631 (WGS 84 / UTM zone 31N).}

\item{additional_cols}{A character vector specifying additional column names
to include in the resulting sf object. Defaults to \code{NULL} (no additional
columns included).}

\item{option}{A character string with "points" (default) for returning sf
points, "lines" to return sf lines and "table" to return a table with a sf
coordinates column or "res_patches" to return sf polygons with residency
patches. For the latter, it is best to specify the buffer around points to
half of \code{lim_spat_indep} of the residency patch calculation. If not
the function can create MULTIPOLYGONS for single residency patches. That will
give a warning message, but works if desired.}

\item{buffer}{A numeric value (in meters) specifying the buffer around the
polygon of each residency patch. This should be set to half of
\code{lim_spat_indep} of the residency patch calculation. If not
the function can create MULTIPOLYGONS for single residency patches. That will
give a warning message, but works if desired.
\code{lim_spat_indep} of the residency patch calculation.}
}
\value{
An \code{sf} object containing the specified coordinates as geometry and
any included attributes.
}
\description{
This function converts a data.frame or data.table to a simple feature (sf)
object, allowing flexible specification of the x and y coordinate columns.
Additional attributes can also be retained in the resulting sf object. There
are four options = c("points", "lines", "table", "res_patches").
}
\examples{
# packages
library(ggplot2)

# load example data
data <- data_example

# example POINTS and LINES

# subset data one tag and tide
data_subset <- data[tag == "3063" & tideID == "2023513"]

# make data spatial
d_sf <- atl_as_sf(
  data_subset,
  additional_cols = c("species", "datetime", "speed_in")
)

# add track
d_sf_lines <- atl_as_sf(
  data_subset,
  additional_cols = c("species", "datetime", "speed_in"),
  option = "lines"
)

# plot interactive map (only if mapview is installed)
if (requireNamespace("mapview", quietly = TRUE)) {
  mapview::mapview(d_sf_lines, zcol = "speed_in", legend = FALSE) +
    mapview::mapview(d_sf, zcol = "speed_in")
}

# example TABLE

# create sf table with spatial points
sf_table <- atl_as_sf(data, x = "x", y = "y", tag = "tag", option = "table")
sf_table

# example RES_PATCHES

# calculate residence patches for one red knot
data <- atl_res_patch(
  data[tag == "3038"],
  max_speed = 3, lim_spat_indep = 75, lim_time_indep = 180,
  min_fixes = 3, min_duration = 120
)

# create polygons around residence patches
d_sf <- atl_as_sf(
  data,
  additional_cols = "patch",
  option = "res_patches", buffer = 75 / 2
)

# summary of residence patches
data_summary <- atl_res_patch_summary(data)

# create basemap
bm <- atl_create_bm(data, buffer = 500)

# geom_sf overwrites coordinate system, so we need to set the limits again
bbox <- atl_bbox(data, buffer = 500)

# plot polygons around residence patches
bm +
  # add patch polygons
  geom_sf(data = d_sf, aes(fill = as.character(patch)), alpha = 0.2) +
  # add track and points
  geom_path(
    data = data, aes(x, y),
    linewidth = 0.1, alpha = 0.5
  ) +
  geom_point(
    data = data[is.na(patch)], aes(x, y),
    size = 0.1, alpha = 0.5, color = "grey20",
    show.legend = FALSE
  ) +
  geom_point(
    data = data[!is.na(patch)], aes(x, y, color = as.character(patch)),
    size = 0.5, show.legend = FALSE
  ) +
  # set extend again (overwritten by geom_sf)
  coord_sf(
    xlim = c(bbox["xmin"], bbox["xmax"]),
    ylim = c(bbox["ymin"], bbox["ymax"]), expand = FALSE
  )
}
\author{
Johannes Krietsch
}
