% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_bbox.R
\name{atl_bbox}
\alias{atl_bbox}
\title{Create a bounding box with specified aspect ratio and buffer}
\usage{
atl_bbox(data, x = "x", y = "y", asp = "16:9", buffer = 0)
}
\arguments{
\item{data}{An \code{sf} or \code{sfc} object for which the bounding box is
calculated or a data.table with x- and y- coordinates.}

\item{x}{A character string representing the name of the column containing
x-coordinates. Defaults to "x".}

\item{y}{A character string representing the name of the column containing
y-coordinates. Defaults to "y".}

\item{asp}{A character string specifying the desired aspect ratio in the
format \code{"width:height"}. Default is \code{"16:9"}.}

\item{buffer}{A numeric value specifying the buffer distance to be applied
to the bounding box. Positive values expand the bounding box, while
negative values shrink it. Default is \code{0}.}
}
\value{
A bounding box (\code{bbox}), represented as a named vector with
\code{xmin}, \code{ymin}, \code{xmax}, and \code{ymax} values.
}
\description{
This function generates a bounding box for a given geometry with a
specified aspect ratio. Additionally, it allows applying a buffer to
expand or contract the bounding box.
}
\examples{
library(sf)

# Create a simple geometry
geom <- st_as_sfc("POLYGON((0 0, 1 0, 1 2, 0 2, 0 0))")

# Create a bounding box with a 16:9 aspect ratio
atl_bbox(geom, asp = "16:9")

# Create a bounding box with a 1:1 aspect ratio and a buffer of 0.5 units
atl_bbox(geom, asp = "1:1", buffer = 0.5)
}
\author{
Johannes Krietsch
}
