% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_check_res_patch.R
\name{atl_check_res_patch}
\alias{atl_check_res_patch}
\title{Check the residence patches from one tag during one tide}
\usage{
atl_check_res_patch(
  data,
  tide_data,
  tide_data_highres,
  tide,
  offset = 0,
  waterlevel_min = -150,
  waterlevel_max = 150,
  waterlevel_line = 60,
  buffer_res_patches,
  buffer_bm = 250,
  buffer_overview = 10000,
  speed_threshold = 3,
  point_size = 1,
  point_alpha = 0.9,
  path_linewidth = 0.5,
  path_alpha = 0.9,
  patch_label_size = 4,
  patch_label_padding = 1,
  patch_alpha = 0.7,
  element_text_size = 11,
  water_fill = "white",
  water_colour = "grey80",
  land_fill = "#faf5ef",
  land_colour = "grey80",
  mudflat_colour = "#faf5ef",
  mudflat_fill = "#faf5ef",
  mudflat_alpha = 0.6,
  roosts = FALSE,
  seed = 1,
  filename = NULL,
  png_width = 3840,
  png_height = 2160
)
}
\arguments{
\item{data}{A \code{data.table} containing tracking data of one tag. Must include
the columns: \code{tag}, \code{x}, \code{y}, \code{time},\code{datetime}, and \code{species} and
\code{patch},  as created by \code{atl_res_patch()}.}

\item{tide_data}{Data on the timing (in UTC) of low and high tides.}

\item{tide_data_highres}{Data on the timing (in UTC) of the waterlevel in
small intervals (e.g. every 10 min) as provided from Rijkwaterstaat.}

\item{tide}{Tide ID to subset.}

\item{offset}{The offset in minutes between the location of the tidal gauge
and the tracking area. This value will be added to the timing of the
water data.}

\item{waterlevel_min}{Numeric in cmNAP. Minimum water level to fix the axis,
if NA (default) it uses the minimum of the provided water level data.
water level to the duration range (default: NA, which uses the minimum of
the provided water level data).}

\item{waterlevel_max}{Numeric in cmNAP. Maximum water level to fix the axis,
if NA (default) it uses the minimum of the provided water level data.}

\item{waterlevel_line}{Numeric in cmNAP. The water level to add as a dotted
line to the plot (default: 60 cmNAP).}

\item{buffer_res_patches}{A numeric value (in meters) specifying the buffer
around the polygon of each residence patch. If set to half of
\code{lim_spat_indep} of the residence patch calculation it reflects the
distance used to determine spatial independence of patches.}

\item{buffer_bm}{Map buffer size (default: 250).}

\item{buffer_overview}{Overview map buffer size (default: 10000).}

\item{speed_threshold}{Speed threshold in m/s for colour scale of movement
speed (default: 3 m/s).}

\item{point_size}{Size of plotted points (default: 1).}

\item{point_alpha}{Transparency of points (default: 0.5).}

\item{path_linewidth}{Line width of movement paths (default: 0.5).}

\item{path_alpha}{Transparency of movement paths (default: 0.2).}

\item{patch_label_size}{Font size for patch labels (default: 4).}

\item{patch_label_padding}{Padding for patch labels (default: 1).}

\item{patch_alpha}{Alpha for patch polygons (default: 0.7).}

\item{element_text_size}{Font size for axis and legend text (default: 11).}

\item{water_fill}{Water fill (default "#D7E7FF")}

\item{water_colour}{Water colour (default "grey80")}

\item{land_fill}{Land fill (default "#faf5ef")}

\item{land_colour}{Land colour (default "grey80")}

\item{mudflat_colour}{Mudflat colour (default "#faf5ef")}

\item{mudflat_fill}{Mudflat fill (default "#faf5ef")}

\item{mudflat_alpha}{Mudflat alpha (default 0.6)}

\item{roosts}{Logical. Whether to add the roost polygon around Griend or not
(default: FALSE).}

\item{seed}{Integer seed used to shuffle the patch colours, so that the
colour order is reproducible. Default is 1.}

\item{filename}{Character (or NULL). If provided, the plot is saved as a
\code{.png} file to this path and with this name; otherwise, the function
returns the plot.}

\item{png_width}{Width of saved PNG (default: 3840).}

\item{png_height}{Height of saved PNG (default: 2160).}
}
\value{
A ggplot object or a saved PNG file.
}
\description{
Generates a \code{ggplot2} showing bird residence patches per tideID, including
movement paths, patch durations, and an inset overview map.
}
\examples{
# packages
library(tools4watlas)

# load example data
data <- data_example

# load example tide pattern and waterlevel data
tidal_pattern_fp <- system.file(
  "extdata", "example-tidalPattern-west_terschelling-UTC.csv",
  package = "tools4watlas"
)
measured_water_height_fp <- system.file(
  "extdata", "example-gemeten_waterhoogte-west_terschelling-clean-UTC.csv",
  package = "tools4watlas"
)
tidal_pattern <- fread(tidal_pattern_fp, yaml = TRUE)
measured_water_height <- fread(measured_water_height_fp)

# calculate residence patches for one red knot
data <- atl_res_patch(
  data[tag == "3038"],
  max_speed = 3, lim_spat_indep = 75, lim_time_indep = 180,
  min_fixes = 2, min_duration = 60
)

# plot example
atl_check_res_patch(
  data[tag == "3038"],
  tide_data = tidal_pattern, tide_data_highres = measured_water_height,
  tide = "2023513", offset = 30,
  buffer_res_patches = 75 / 2
)
}
\author{
Johannes Krietsch
}
