% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_check_tag.R
\name{atl_check_tag}
\alias{atl_check_tag}
\title{Check the data from one tag on a map}
\usage{
atl_check_tag(
  data,
  buffer = 1000,
  asp = "16:9",
  option = "datetime",
  viridis_option = "A",
  viridis_direction = -1,
  first_n = NULL,
  last_n = NULL,
  highlight_first = FALSE,
  highlight_last = FALSE,
  point_size = 0.5,
  point_alpha = 1,
  path_linewidth = 0.5,
  path_alpha = 0.1,
  filename = NULL
)
}
\arguments{
\item{data}{A \code{data.table} containing tracking data. Must include the
columns: \code{"tag"}, \code{"x"}, \code{"y"}, \code{"time"}, and \code{"datetime"}.}

\item{buffer}{Numeric. The buffer size in meters around the data points in
the plot (default: 1000).}

\item{asp}{Character. The aspect ratio of the plot (default: \code{"16:9"}).}

\item{option}{Character. Determines the color mapping variable. Options are:
\itemize{
\item \code{"datetime"}:
\item \code{"nbs"}:
\item \code{"sd"}:
\item \code{"speed_in"}:
\item \code{"gap"}:
}}

\item{viridis_option}{Character. The color scheme option from \code{viridis}
(default: \code{"A"}). See
https://search.r-project.org/CRAN/refmans/viridisLite/html/viridis.html
for all options (A-H).}

\item{viridis_direction}{Numeric. Direction of the viridis color scale
(-1 reverses, default: -1).}

\item{first_n}{Numeric (or NULL). If provided, only the first \code{n} locations
are shown.}

\item{last_n}{Numeric (or NULL). If provided, only the last \code{n} locations
are shown.}

\item{highlight_first}{Logical. If \code{TRUE}, highlights the first point in the
track (default: \code{FALSE}).}

\item{highlight_last}{Logical. If \code{TRUE}, highlights the last point in the
track (default: \code{FALSE}).}

\item{point_size}{Numeric. The size of the data points (default: 0.5).}

\item{point_alpha}{Numeric. Transparency of the data points (default: 1).}

\item{path_linewidth}{Numeric. The width of the connecting track lines
(default: 0.5).}

\item{path_alpha}{Numeric. Transparency of the track lines (default: 0.1).}

\item{filename}{Character (or NULL). If provided, the plot is saved as a
\code{.png} file to this path and with this name; otherwise, the function
returns the plot.}
}
\value{
A \code{ggplot2} object with the specified option and adjustments. If
\code{filename} is provided, the plot is saved as a \code{.png} file instead of
being returned.
}
\description{
This function processes tracking data for a specific tag and generates a
visualization using \code{ggplot2}. It allows customization of colors, point
sizes, and track styles, and supports various display options such as
datetime, nbs (number of base stations / receivers), standard deviation,
speed_in and gap. The function can either return the plot or save it as an
png file.
}
\examples{
# packages
library(tools4watlas)

# path to csv with filtered data
data_path <- system.file(
  "extdata", "watlas_data_filtered.csv",
  package = "tools4watlas"
)

# load data
data <- fread(data_path, yaml = TRUE)

# subset bar-tailed godwit
data <- data[species == "bar-tailed godwit"]

# plot different options
atl_check_tag(
  data,
  option = "datetime",
  highlight_first = TRUE, highlight_last = TRUE
)
atl_check_tag(data, option = "nbs")
atl_check_tag(data, option = "sd")
atl_check_tag(data, option = "speed_in")
atl_check_tag(data, option = "gap")
}
