% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_create_bm.R
\name{atl_create_bm}
\alias{atl_create_bm}
\title{Create a base map with customised bounding box}
\usage{
atl_create_bm(
  data = NULL,
  x = "x",
  y = "y",
  buffer = 100,
  asp = "16:9",
  option = "osm",
  land_data = tools4watlas::land,
  mudflats_data = tools4watlas::mudflats,
  lakes_data = tools4watlas::lakes,
  raster_data,
  sc_dist,
  sc_location = "br",
  sc_cex = 0.7,
  sc_height = unit(0.25, "cm"),
  sc_pad_x = unit(0.25, "cm"),
  sc_pad_y = unit(0.5, "cm"),
  projection = sf::st_crs(32631)
)
}
\arguments{
\item{data}{A \code{data.table} or an object convertible to \code{data.table}
containing spatial points. Defaults to a single point around Griend if
\code{NULL}.}

\item{x}{A character string specifying the column with x-coordinates.
Defaults to \code{"x"}.}

\item{y}{A character string specifying the column with y-coordinates.
Defaults to \code{"y"}.}

\item{buffer}{A numeric value specifying the buffer distance for the bounding
box. Default is \code{1000}.}

\item{asp}{A character string specifying the aspect ratio in \code{"width:height"}
format. Default is \code{"16:9"}.}

\item{option}{Either "osm" for OpenStreetMap polygons or "bathymetry" for
bathymetry data. Note that for the later it is necessary to provide the
bathymetry data in UTM31.}

\item{land_data}{An \code{sf} object for land polygons. Defaults to \code{land}.}

\item{mudflats_data}{An \code{sf} object for mudflat polygons. Defaults to
\code{mudflats}.}

\item{lakes_data}{An \code{sf} object for lake polygons. Defaults to \code{lakes}.}

\item{raster_data}{An \code{SpatRaster} (tif opened with \code{terra::rast()} of
bathymetry data.}

\item{sc_dist}{Scale bar distance. Optional; calculated automatically if
omitted.}

\item{sc_location}{A character string specifying the location of the scale
bar. Default is \code{"br"} (bottom right).}

\item{sc_cex}{Numeric value for the scale bar text size. Default is \code{0.7}.}

\item{sc_height}{A unit object specifying the height of the scale bar.
Default is \code{unit(0.25, "cm")}.}

\item{sc_pad_x}{A unit object specifying horizontal padding for the scale
bar. Default is \code{unit(0.25, "cm")}.}

\item{sc_pad_y}{A unit object specifying vertical padding for the scale bar.
Default is \code{unit(0.5, "cm")}.}

\item{projection}{The coordinate reference system (CRS) for the spatial data.
Defaults to EPSG:32631 (WGS 84 / UTM zone 31N). Output is always UTM 31N}
}
\value{
A \code{ggplot2} object representing the base map with the specified
settings.
}
\description{
This function creates a base map using spatial data layers, allowing for
custom bounding boxes, aspect ratios, and scale bar adjustments.
}
\examples{
# Example with default settings (map around Griend)
bm <- atl_create_bm(buffer = 5000)
print(bm)
}
\author{
Johannes Krietsch
}
