% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_create_bm_tiles.R
\name{atl_create_bm_tiles}
\alias{atl_create_bm_tiles}
\title{Create a basemap with customised bounding box using map tiles}
\usage{
atl_create_bm_tiles(
  data = NULL,
  x = "x",
  y = "y",
  buffer = 100,
  asp = "16:9",
  option = "Esri.WorldImagery",
  zoom = 15,
  scalebar = TRUE,
  sc_location = "br",
  sc_cex = 1,
  sc_height = 0.3,
  sc_pad_x = 0.4,
  sc_pad_y = 0.6,
  projection = sf::st_crs(32631)
)
}
\arguments{
\item{data}{A \code{data.table} or an object convertible to \code{data.table}
containing spatial points or a \code{sf} bounding box.
Defaults to a single point around Griend if \code{NULL}.}

\item{x}{A character string specifying the column with x-coordinates.
Defaults to \code{"x"}.}

\item{y}{A character string specifying the column with y-coordinates.
Defaults to \code{"y"}.}

\item{buffer}{A numeric value (in meters) specifying the buffer distance for
the bounding box. Default is \code{1000}.}

\item{asp}{A character string specifying the desired aspect ratio in the
format \code{"width:height"}. Default is \code{"16:9"}, if \code{NULL} returns simple
bounding box without modifying aspect ratio.}

\item{option}{A character string specifying the map tile provider.
Options include \code{"Esri.WorldImagery"}, \verb{""OpenStreetMap"},  \code{"Esri"},
\code{"CARTO"}, and \code{"Thunderforest"}. See
supported by the \code{maptiles} package, see:}

\item{zoom}{Numeric value specifying the zoom level for the map tiles.
Zoom levels are described in the OpenStreetMap wiki:
https://wiki.openstreetmap.org/wiki/Zoom_levels.}

\item{scalebar}{TRUE or FALSE for adding a scalebar to the plot.}

\item{sc_location}{A character string specifying the location of the scale
bar. Default is \code{"br"} (bottom right).}

\item{sc_cex}{Numeric value for the scale bar text size. Default is \code{0.7}.}

\item{sc_height}{A unit object specifying the height of the scale bar.
Default is \code{unit(0.25, "cm")}.}

\item{sc_pad_x}{A unit object specifying horizontal padding for the scale
bar. Default is \code{unit(0.25, "cm")}.}

\item{sc_pad_y}{A unit object specifying vertical padding for the scale bar.
Default is \code{unit(0.5, "cm")}.}

\item{projection}{The coordinate reference system (CRS) for the spatial data.
Defaults to EPSG:32631 (WGS 84 / UTM zone 31N). Output is always EPSG:4326.
Bounding box calculation is much faster when uusing EPSG:3263, so use it
like this whenwever possible and then only plot the movement tracks in
EPSG:4326 on the map.}
}
\value{
A \code{ggplot2} object representing the base map with the specified
settings.
}
\description{
This function creates a basemap using spatial data layers, allowing for
custom bounding boxes, aspect ratios, and scale bar adjustments.
}
\examples{
# packages
library(tools4watlas)
library(ggplot2)

# example with open street map
bm <- atl_create_bm_tiles(
  buffer = 15000, option = "OpenStreetMap", zoom = 12
)
print(bm)

# example with satellite map
bm <- atl_create_bm_tiles(
  buffer = 15000, option = "Esri.WorldImagery", zoom = 12
)
print(bm)

# example with bbox from data and movement data
data <- data_example

# add transformed coordinates in projection of the base map (EPSG:4326)
data <- atl_transform_dt(data)

# plot points and tracks using transformed coordinates.
bm +
  geom_path(
    data = data, aes(x_4326, y_4326, colour = tag),
    linewidth = 0.5, alpha = 0.1, show.legend = FALSE
  ) +
  geom_point(
    data = data, aes(x_4326, y_4326, colour = tag),
    size = 0.5, alpha = 1, show.legend = FALSE
  ) +
  scale_color_discrete(name = paste("N = ", length(unique(data$tag)))) +
  theme(legend.position = "top")
}
\author{
Johannes Krietsch
}
