% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_file_path.R
\name{atl_file_path}
\alias{atl_file_path}
\title{Get the file path for WATLAS or GIS data based on the user's name.}
\usage{
atl_file_path(data_type = c("watlas_teams", "rasters", "shapefiles"))
}
\arguments{
\item{data_type}{A character string indicating the type of data. Options are:
\itemize{
\item \code{"watlas_teams"}: Path to WATLAS team data.
\item \code{"rasters"}: Path to raster GIS data.
\item \code{"shapefiles"}: Path to shapefile GIS data.
}}
}
\value{
A character string representing the full file path to the selected
data type for the current user.
}
\description{
This function returns a predefined file path based on the user's system
username and the selected data type.
Users must be predefined in the function.
}
\examples{
atl_file_path("watlas_teams")
atl_file_path("rasters")
atl_file_path("shapefiles")
}
