% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_filter_covariates.R
\name{atl_filter_covariates}
\alias{atl_filter_covariates}
\title{Filter data by position covariates}
\usage{
atl_filter_covariates(data, filters = c(), quietly = FALSE)
}
\arguments{
\item{data}{A \code{data.table} or similar containing the variables to be
filtered.}

\item{filters}{A character vector of filter expressions. An example might be
\code{"speed < 20"}. The filtering variables must be in the \code{data.table}.
The function will not explicitly check whether the filtering variables are
present; this makes it flexible, allowing expressions such as
\code{"between(speed, 2, 20)"}, but also something to use at your own risk.
A missing filter variables \emph{will} result in an empty data frame.}

\item{quietly}{If \code{TRUE} returns percentage and number of positions filtered,
if \code{FALSE} functions runs quietly}
}
\value{
A dataframe filtered using the filters specified.
}
\description{
The atlastools function \code{atl_filter_covariates} allows convenient
filtering of a dataset by any number of logical filters.
This function can be used to easily filter timestamps in a range, as well as
combine simple spatial and temporal filters.
It accepts a character vector of \code{R} expressions that each return a
logical vector (i.e. \code{TRUE} or \code{FALSE}).
Each filtering condition is interpreted in the context of the dataset
supplied, and used to filter for rows that satisfy each of the filter
conditions.
Users must make sure that the filtering variables exist in their dataset in
order to avoid errors.
}
\examples{
# packages
library(tools4watlas)

# load example data
data <- data_example

# filter data at night
# extract hour of the day
data[, hour := as.integer(format(datetime, "\%H"))]

night_data <- atl_filter_covariates(
  data = data,
  filters = c("!inrange(hour, 6, 18)")
)

# filter on the variance of the estimated x- and y-coordinates
var_max <- 5000 # in meters squared

data_filtered <- atl_filter_covariates(
  data = data,
  filters = c(
    sprintf("varx < \%s", var_max),
    sprintf("vary < \%s", var_max)
  )
)

# filter by speed
speed_max <- 35 # m/s (126 km/h)

data <- atl_filter_covariates(
  data = data,
  filters = c(
    sprintf("speed_in < \%s | is.na(speed_in)", speed_max),
    sprintf("speed_out < \%s | is.na(speed_out)", speed_max)
  )
)
}
\author{
Pratik R. Gupte and Johannes Krietsch
}
