% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxiliary.r
\name{atl_format_time}
\alias{atl_format_time}
\title{Format time in easy readable interval}
\usage{
atl_format_time(time)
}
\arguments{
\item{time}{Time in seconds (numeric or vector of numeric values).}
}
\value{
A character vector with the formatted time intervals.
}
\description{
This function converts a given time (in seconds) into a easy readable format
with days, hours, minutes, or seconds.
}
\examples{
library(tools4watlas)
atl_format_time(3600)
atl_format_time(c(120, 3600, 86400))
}
\author{
Johannes Krietsch
}
