% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_data.R
\name{atl_get_data}
\alias{atl_get_data}
\title{Get data from a SQLite-database.}
\usage{
atl_get_data(
  tag,
  tracking_time_start,
  tracking_time_end,
  timezone = "CET",
  database = "some_database",
  host = "some.host.nl",
  username = "someuser",
  password = "somepassword",
  SQLiteDB = NULL,
  use_connection = NULL
)
}
\arguments{
\item{tag}{An 11 digit number representing the ATLAS tag. May be passed as a character or a numeric, and will be converted to a character. For example \code{"31001000001"}.}

\item{tracking_time_start}{Character representation of time from which start-point data should be retrieved.}

\item{tracking_time_end}{Character time representing the end point corresponding to the above start point.}

\item{timezone}{Timezone of tracking_time_start and tracking_time_end. Must be a valid timezone for the as.POSIXct function.}

\item{database}{The database name on the host server.}

\item{host}{The server address on which the data are stored.}

\item{username}{Username to access the server.}

\item{password}{Password to access the server.}

\item{SQLiteDB}{The full path including the filename of a local SQLite-database whith tracking data. This argument overrules establishing a connection with the remote host.}

\item{use_connection}{An R-object with a previously established connection, for instance generated by \code{RMySQL::dbConnect(RMySQL::MySQL(), user = 'username', password = 'password', dbname = 'database', host = 'host')}. This argument overrules making a connection to a local host \code{SQLiteDB} or a remote host.}
}
\value{
A dataframe of localizations of the specified tag, filtered between the start and end times.\cr
PosID	=	Unique number for localizations \cr
TAG 		=	11 digit WATLAS tag ID \cr
tag		=	4 digit tag number (character) \cr
TIME		=	UNIX time (seconds) \cr
time 	= 	Timestamp in POSIXct (UTC) \cr
X		=	X-ccordinates in meters (utm 31 N) \cr
Y		=	Y-ccordinates in meters (utm 31 N) \cr
NBS		=	Number of Base Stations used in calculating coordinates \cr
VARX		=	Variance in estimating X-coordinates \cr
VARY		=	Variance in estimating Y-coordinates \cr
COVXY	=	Co-variance between X- and Y-coordinates \cr
}
\description{
this function can be used to obtain data from a local or remote SQLite database. If the full path with filename is provided in 'SQLiteDB', it will look for that SQLite-file locally. If 'SQLiteDB' is NULL it will establish a connection with a remote server. There is also functionality to use an existing opened data base connection that can be provided in 'use_connection'. For instance, when creating residence patches within a loop of tags and tide ID's. Because establishing a conncection within a for-loop takes a long time, an established connection can be used.
}
\author{
Pratik R. Gupte & Allert I. Bijleveld
}
