% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_data.R
\name{atl_get_data}
\alias{atl_get_data}
\title{Get data from a SQLite-database}
\usage{
atl_get_data(
  tag,
  tracking_time_start,
  tracking_time_end,
  timezone = "CET",
  database = "some_database",
  host = "some.host.nl",
  username = "someuser",
  password = "somepassword",
  SQLiteDB = NULL,
  use_connection = NULL
)
}
\arguments{
\item{tag}{An number representing the WATLAS tag (either in short format
e.g. 4 digits or long format with 11 digits). May be passed as
a character or a numeric, and will be converted to a character.
For example \code{"31001000001"}.}

\item{tracking_time_start}{Character representation of time from which
start-point data should be retrieved.}

\item{tracking_time_end}{Character time representing the end point
corresponding to the above start point.}

\item{timezone}{Timezone of tracking_time_start and tracking_time_end.
Must be a valid timezone for the as.POSIXct function.}

\item{database}{The database name on the host server.}

\item{host}{The server address on which the data are stored.}

\item{username}{Username to access the server.}

\item{password}{Password to access the server.}

\item{SQLiteDB}{The full path including the filename of a local
SQLite-database with tracking data. This argument overrules establishing
a connection with the remote host.}

\item{use_connection}{An R-object with a previously established connection,
for instance generated by \code{RMySQL::dbConnect(RMySQL::MySQL(),
 user = 'username', password = 'password', dbname = 'database',
 host = 'host')}. This argument overrules making a connection to a
local host \code{SQLiteDB} or a remote host.}
}
\value{
A data.table of positions of the specified tag, filtered between
the start and end times.\cr
posID	  =	Unique number for positions \cr
tag		  =	4 digit tag number (character) \cr
time		  =	UNIX time (seconds) \cr
datetime = Time stamp in POSIXct (UTC) \cr
x		    =	x-coordinates in meters (UTM 31 N) \cr
y		    =	y-coordinates in meters (UTM 31 N) \cr
nbs		  =	Number of Base Stations (receivers) used in calculating
coordinates \cr
varx		  =	Variance in estimating x-coordinates \cr
vary		  =	Variance in estimating y-coordinates \cr
covxy	  =	Co-variance between y- and y-coordinates \cr
}
\description{
This function can be used to obtain data from a local or remote SQLite
database. If the full path with filename is provided in \code{SQLiteDB}, it will
look for that SQLite-file locally. If \code{SQLiteDB} is \code{NULL} it will establish
a connection with a remote server. There is also functionality to use an
existing opened data base connection that can be provided in
\code{use_connection}. For instance, when creating residence patches within a
loop of tags and tide ID's. Because establishing a connection within a
for-loop takes a long time, an established connection can be used.
}
\author{
Pratik R. Gupte & Allert I. Bijleveld & Johannes Krietsch
}
