% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_speed.R
\name{atl_get_speed}
\alias{atl_get_speed}
\title{Calculate instantaneous speed}
\usage{
atl_get_speed(
  data,
  tag = "tag",
  x = "x",
  y = "y",
  time = "time",
  type = c("in", "out")
)
}
\arguments{
\item{data}{A dataframe or similar which must have the columns
specified by \code{x}, \code{y}, and \code{time}.}

\item{tag}{The tag ID.}

\item{x}{The x coordinate.}

\item{y}{The y coordinate.}

\item{time}{The timestamp in seconds since the UNIX epoch.}

\item{type}{The type of speed (incoming or outgoing) to return.
Incoming speeds are specified by \code{type = "in"}, and outgoing speeds
by \code{type = "out"} or both c("in", "out").}
}
\value{
Data.table changed in place with additional speed columns
}
\description{
Returns additional columns for incoming and outcoming speed to the
data.table. Speed in metres per time interval. The time interval is dependent
on the units of the column specified in \code{time}.
}
\details{
Derived from \code{atlastools::atl_get_speed()} in the \pkg{atlastools}
package (Gupte et al., 2022), licensed under GPL-3.
}
\examples{
# packages
library(tools4watlas)

# load example data
data <- data_example

# remove speed columns
data[, c("speed_in", "speed_out") := NULL]

# calculate speed
data <- atl_get_speed(data,
                      tag = "tag",
                      x = "x",
                      y = "y",
                      time = "time",
                      type = c("in", "out")
)

# check data
data[, .(tag, datetime, x, y, speed_in, speed_out)]
}
\references{
Gupte, P. R., Beardsworth, C. E., Spiegel, O., Lourie, E., Toledo, S.,
Nathan, R., & Bijleveld, A. I. (2022). A guide to pre-processing
high-throughput animal tracking data. \emph{Journal of Animal Ecology}, 91,
287-307. \doi{10.1111/1365-2656.13610}
}
\author{
Pratik R. Gupte, Allert Bijleveld & Johannes Krietsch
}
