% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_smooth_data.R
\name{atl_median_smooth}
\alias{atl_median_smooth}
\title{Apply a median smooth to coordinates.}
\usage{
atl_median_smooth(data, x = "x", y = "y", time = "time", moving_window = 5)
}
\arguments{
\item{data}{A dataframe object returned by \code{atl_get_data}, which should contain the original columns (particularly X,Y and TIME).}

\item{x}{The X coordinate.}

\item{y}{The Y coordinate.}

\item{time}{The timestamp, ideally as an integer.}

\item{moving_window}{The size of the moving window for the median smooth. Must be an odd number.}
}
\value{
A datatable class object (extends data.frame), including X,Y as smoothed coordinates and the x_raw and y_raw, which are the raw coordinates.
}
\description{
Applies a median smooth defined by a rolling window to the X and Y
coordinates of the data.
}
\examples{
\dontrun{
atl_median_smooth(
  data = track_data,
  x = "x", y = "y",
  time = "time",
  moving_window = 5
)
}

}
\author{
Pratik Gupte & Allert Bijleveld
}
