% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot_tag.R
\name{atl_plot_tag}
\alias{atl_plot_tag}
\title{Plot track for one individual on a simple background.}
\usage{
atl_plot_tag(
  data,
  tag = NULL,
  fullname = NULL,
  color_by = "time",
  towers = NULL,
  h = 7,
  w = 7 * (16/9),
  buffer = 1,
  legend = "topleft",
  scalebar = 5,
  cex_legend = 1,
  land_data = tools4watlas::land_sf,
  mudflats_data = tools4watlas::mudflats_sf,
  lakes_data = tools4watlas::lakes_sf,
  rivers_data = tools4watlas::rivers_sf
)
}
\arguments{
\item{data}{A dataframe with the tracking data. Can include multiple tags,
but one tag is selected for plotting.}

\item{tag}{The four-digit tag number as character to plot. Defaults to
plotting the first tag in \code{data}.}

\item{fullname}{If specified the plot will be saved in this path with this
name (include extension). Defaults to NULL and plotting in a graphics window.}

\item{color_by}{Either \code{"time"}, \code{"sd"}, or \code{"nbs"},
which are respectively used to colour the localization with the relative
time (hours), variance in the localizations as the maximum of VARX and
VARY, or the Number of Base Stations (NBS) used to calculate the
localization. Defaults to "time".}

\item{towers}{A dataframe with coordinates of receiver stations
(named \code{X} and \code{Y}).}

\item{h}{height of the plot (when saving)}

\item{w}{width of the plot (when saving)}

\item{buffer}{Buffer around bounding box in meters}

\item{legend}{Passed to the \code{legend} function and sets the location
of the legend in the plot.}

\item{scalebar}{Length of scalebar in km.}

\item{cex_legend}{The size of the text in the legend.}

\item{land_data}{An \code{sf} object for land polygons. Defaults to \code{land_sf}.}

\item{mudflats_data}{An \code{sf} object for mudflat polygons. Defaults to
\code{mudflats_sf}.}

\item{lakes_data}{An \code{sf} object for lake polygons. Defaults to \code{lakes_sf}.}

\item{rivers_data}{An \code{sf} object for river polygons. Defaults to
\code{rivers_sf}.}
}
\value{
Returns nothing but a plot.
}
\description{
A function that plots the localization data of one individual.
}
\examples{
library(tools4watlas)

# Load example data
data <- data_example

# Transform to sf
d_sf <- atl_as_sf(data, additional_cols = names(data))

# Plot the tracking data with a simple background
atl_plot_tag(
  data = d_sf, tag = NULL, fullname = NULL, buffer = 1,
  color_by = "time"
)

atl_plot_tag(
  data = d_sf, tag = NULL, fullname = NULL, buffer = 1,
  color_by = "sd"
)

atl_plot_tag(
  data = data, tag = NULL, fullname = NULL, buffer = 1,
  color_by = "time"
)
}
\author{
Allert Bijleveld & Johannes Krietsch
}
