% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_animate_data.R
\name{atl_progress_bar}
\alias{atl_progress_bar}
\title{Display a live progress bar for PNG file generation in a directory}
\usage{
atl_progress_bar(file_path, total = NULL, refresh_rate = 1)
}
\arguments{
\item{file_path}{Path to the directory containing PNG files.}

\item{total}{Optional. Total number of expected PNG files. If NULL, the
function reads from 'total_frames.txt' created by atl_time_steps() in the
same directory.}

\item{refresh_rate}{Numeric value in seconds specifying how often the
progress bar updates.}
}
\value{
No return value. Prints progress bar to the console.
}
\description{
Display a live progress bar for PNG file generation in a directory
}
