% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_make_res_patch.r
\name{atl_res_patch}
\alias{atl_res_patch}
\title{Construct residence patches from position data}
\usage{
atl_res_patch(
  data,
  max_speed = 3,
  lim_spat_indep = 75,
  lim_time_indep = 180,
  min_fixes = 3,
  min_duration = 120,
  summary_variables = c(),
  summary_functions = c()
)
}
\arguments{
\item{data}{A dataframe of any class that is or extends data.frame of one
individual only. The dataframe must contain at least two spatial coordinates,
\code{x} and \code{y}, and a temporal coordinate, \code{time}.}

\item{max_speed}{A numeric value specifying the maximum speed (m/s) between
two coordinates that would be considered non-transitory}

\item{lim_spat_indep}{A numeric value of distance in metres of the spatial
distance between two patches for them to the considered independent.}

\item{lim_time_indep}{A numeric value of time in minutes of the time
difference between two patches for them to be considered independent.}

\item{min_fixes}{The minimum number of fixes for a group of
spatially-proximate number of ponts to be considered a preliminary residence
patch.}

\item{min_duration}{The minimum duration (in seconds) for classifying
residence patches.}
}
\value{
A data.table that has the added column
\code{patch} indicating the patch.
}
\description{
A cleaned movement track of one individual at a time can be classified into
residence patches using the
function \code{atl_res_patch}.
The function expects a specific organisation of the data: there should be
at least the following columns, \code{x}, \code{y}, and \code{time},
corresponding to the coordinates, and the time as \code{POSIXct}.
\code{atl_res_patch} requires only three parameters: (1) the maximum
speed threshold between localizations (called \code{max_speed}), (2) the
distance threshold between clusters of positions (called
\code{lim_spat_indep}), and (3) the time interval between clusters
(called \code{lim_time_indep}).Clusters formed of fewer than a minimum
number of positions can be excluded.The exclusion of clusters with few
positions can help in removing bias due to short stops, but if such short
stops are also of interest, they can be included by reducing the
\code{min_fixes} argument.
}
\author{
Pratik R. Gupte, Christine E. Beardsworth & Allert I. Bijleveld &
Johannes Krietsch
}
