% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxiliary.r
\name{atl_spec_cols}
\alias{atl_spec_cols}
\title{WATLAS species colours}
\usage{
atl_spec_cols(option = "vector")
}
\arguments{
\item{option}{A character string specifying the output format. Options are
\code{"vector"} (default), which returns a named colour vector, or \code{"table"},
which returns a data.table with species names and colours.}
}
\value{
A named character vector (for use in ggplot2) or a data.table
with species names and corresponding colours.
}
\description{
Returns a vector or table of predefined colours for WATLAS species.
}
\examples{
library(tools4watlas)
atl_spec_cols("vector")
atl_spec_cols("table")
}
