% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxiliary.r
\name{atl_t_col}
\alias{atl_t_col}
\title{Make a colour transparant.}
\usage{
atl_t_col(color, percent = 50, name = NULL)
}
\arguments{
\item{color}{The color to make transparant.}

\item{percent}{The percentage of transparancy to apply .}

\item{name}{The name argument as passed on to rgb.}
}
\value{
The transparant color will be returned.
}
\description{
A functionm that will make the provided colour transparant.
}
\author{
Allert Bijleveld
}
