% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxiliary.R
\name{atl_t_col}
\alias{atl_t_col}
\title{Make a colour transparant.}
\usage{
atl_t_col(color, percent = 50, name = NULL)
}
\arguments{
\item{color}{The color to make transparant.}

\item{percent}{The percentage of transparancy to apply .}

\item{name}{The name argument as passed on to rgb.}
}
\value{
The transparant color will be returned.
}
\description{
A functionm that will make the provided colour transparant.
}
\examples{
# Example with 50\% transparency
color_with_alpha <- atl_t_col("blue", percent = 50)
print(color_with_alpha)

plot(1, 1,
  col = color_with_alpha, pch = 16, cex = 20,
  xlab = "X", ylab = "Y", main = "Point with Transparent Color"
)

# Example with 30\% transparency
color_with_alpha <- atl_t_col("red", percent = 90)
print(color_with_alpha)

plot(1, 1,
  col = color_with_alpha, pch = 16, cex = 20,
  xlab = "X", ylab = "Y", main = "Point with Transparent Color"
)
}
\author{
Allert Bijleveld & Johannes Krietsch
}
