% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxiliary.R
\name{atl_t_col}
\alias{atl_t_col}
\title{Make a colour transparent}
\usage{
atl_t_col(color, percent = 50, name = NULL)
}
\arguments{
\item{color}{The colour to make transparent.}

\item{percent}{The percentage of transparency to apply .}

\item{name}{The name argument as passed on to rgb.}
}
\value{
The transparent colour will be returned.
}
\description{
A function that will make the provided colour transparent.
}
\examples{
# Example with 50\% transparency
colour_with_alpha <- atl_t_col("blue", percent = 50)
print(colour_with_alpha)

plot(1, 1,
  col = colour_with_alpha, pch = 16, cex = 20,
  xlab = "X", ylab = "Y", main = "Point with Transparent Colour"
)

# Example with 30\% transparency
colour_with_alpha <- atl_t_col("red", percent = 90)
print(colour_with_alpha)

plot(1, 1,
  col = colour_with_alpha, pch = 16, cex = 20,
  xlab = "X", ylab = "Y", main = "Point with Transparent Colour"
)
}
\author{
Allert Bijleveld & Johannes Krietsch
}
