% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxiliary.r
\name{atl_tag_cols}
\alias{atl_tag_cols}
\title{Assign colours to tag ID's}
\usage{
atl_tag_cols(tags, option = "vector")
}
\arguments{
\item{tags}{A character or numeric vector of tags. Duplicate values are
ignored.}

\item{option}{A string indicating the output format. Either \code{"vector"}
for a named character vector of hex colors, or \code{"table"} for a
\code{data.table} with columns \code{tag} and \code{colour}. Default is
\code{"vector"}.}
}
\value{
A named character vector of hex color codes if
\code{option = "vector"}, or a \code{data.table} with two columns
(\code{tag}, \code{colour}) if
\code{option = "table"}.
}
\description{
Generates distinct and visually distinct colours for a set of input tags
using the default hue-based palette from ggplot2
(via \code{scales::hue_pal()}).
This is useful for assigning consistent colours to categorical labels in
plots, for example in an animation, where not all individuals always have
data in each frame.
}
\examples{
# Default output (named vector)
atl_tag_cols(c("1234", "2121", "9999"))

# Output as a data.table
atl_tag_cols(c("1234", "2121", "9999"), option = "table")

}
