% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_auxiliary.r
\name{atl_tag_labs}
\alias{atl_tag_labs}
\title{Create unique labels for tags by combining specified columns}
\usage{
atl_tag_labs(data, columns, sep = " ")
}
\arguments{
\item{data}{A \code{data.table} or \code{data.frame} containing at least a
\code{tag} column.}

\item{columns}{A character vector specifying the names of columns to combine
into the label. These columns must exist in \code{data}.}

\item{sep}{A character string used to separate concatenated column values in
the label. Defaults to an space \code{" "}.}
}
\value{
A named character vector where the names are unique \code{tag} values
from the data, and the values are concatenated labels created from the
specified columns.
}
\description{
Generates a named character vector of unique labels for each \code{tag} in the
data. The labels are created by concatenating the values of specified columns
for the first occurrence of each tag, using a user-defined separator.
}
\details{
The function selects the first row for each unique \code{tag} to create the
label. If the \code{tag} column or any of the specified columns are missing,
the function will stop with an informative error message.
}
\examples{
library(data.table)
data <- data.table(
  tag = c("1234", "2222", "1234"),
  rings = c(123234442, 334234234, 123234442),
  name = c("Allert", "Peter", "Karl")
)
atl_tag_labs(data, c("rings", "name"), sep = " ")

}
