% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_angle.R
\name{atl_turning_angle}
\alias{atl_turning_angle}
\title{Get the turning angle between points.}
\usage{
atl_turning_angle(data, X = "X", Y = "Y", time = "TIME")
}
\arguments{
\item{data}{A dataframe or similar ordered by time, which must have the columns
specified by \code{X}, \code{Y}, and \code{TIME}.}

\item{X}{The X coordinate.}

\item{Y}{The Y coordinate.}

\item{time}{The timestamp in seconds since the UNIX epoch.}
}
\value{
A vector of turning angles in degrees.
Negative degrees indicate 'left' turns. There are two fewer
angles than the number of rows in the dataframe.
}
\description{
Gets the relative heading between two track segments (three localizations) using the law of cosines.
The turning angle is returned in degrees.
Users should apply this function to \emph{one individual at a time}, ideally by
splittng a dataframe with multiple individuals into a list of dataframes.
}
\examples{
\dontrun{
data$angle <- atl_turning_angle(data,
  X = "X", Y = "Y", time = "TIME"
)
}
}
\author{
Pratik R. Gupte & Allert Bijleveld
}
