% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_within_polygon.r
\name{atl_within_polygon}
\alias{atl_within_polygon}
\title{Detect position intersections with a polygon}
\usage{
atl_within_polygon(
  data,
  x = "x",
  y = "y",
  polygon,
  col_name = deparse(substitute(polygon))
)
}
\arguments{
\item{data}{A \code{data.table} or similar containing at least x and
y coordinates.}

\item{x}{The name of the x coordinate, default "x".}

\item{y}{The name of the y coordinate, default "y".}

\item{polygon}{An \code{sf} polygon object with a EPSG:32631 (UTM zone 31N) as
CRS.}

\item{col_name}{The name of the output column added to \code{data}. Defaults
to the name of the polygon object passed in.}
}
\value{
The original \code{data} with an added logical column indicating
whether each position intersects the polygon.
}
\description{
Detects which positions intersect a polygon sf object.
}
\examples{
# packages
library(tools4watlas)
library(sf)
library(ggplot2)

# load example data
data <- data_example

# create basemap
bm <- atl_create_bm(data, buffer = 800)

# create a bounding box to filter data
griend_east <- st_sfc(st_point(c(5.275, 53.2523)), crs = st_crs(4326)) |>
  st_transform(crs = st_crs(32631))

# define bbox to crop data
bbox_crop <- atl_bbox(griend_east, asp = "16:9", buffer = 2000)
bbox_sf <- st_as_sfc(bbox_crop) # just for plotting as sf object

# geom_sf overwrites coordinate system, so we need to set the limits again
bbox <- atl_bbox(data, buffer = 800)


data <- atl_within_polygon(data, polygon = bbox_sf)
}
\author{
Johannes Krietsch
}
