% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_example}
\alias{data_example}
\title{Data from two red knots and one redshank}
\format{
data.table of watlas data with tide data added
}
\source{
watlas data example
}
\usage{
data_example
}
\description{
Data from two red knots and one redshank
}
\keyword{datasets}
