% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{grienderwaard}
\alias{grienderwaard}
\title{Grienderwaard}
\format{
sf object of the Grienderwaard
}
\source{
Manually created based on openstreetmap data and -161 cm NAP polygon.
}
\usage{
grienderwaard
}
\description{
Grienderwaard
}
\keyword{datasets}
