% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lakes}
\alias{lakes}
\title{Lake on Griend}
\format{
sf object used for background layer in the function atl_create_bm()
}
\source{
\url{https://www.openstreetmap.org}
}
\usage{
lakes
}
\description{
Lake on Griend
}
\keyword{datasets}
