% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land}
\alias{land}
\title{The Netherlands}
\format{
\describe{
\item{SOORT}{land, water, sea (zee)}
}
}
\source{
\url{https://www.rijkswaterstaat.nl}
}
\usage{
land
}
\description{
SpatialPolygonsDataFrame describing the Netherlands' surface areas. Used for
background layer in the function atl_plot_tag.
}
\keyword{datasets}
