% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mudflats}
\alias{mudflats}
\title{Intertidal Dutch Wadden Sea}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 175 rows and 3 columns.
}
\source{
\url{https://www.rijkswaterstaat.nl}
}
\usage{
mudflats
}
\description{
SpatialPolygonsDataFrame describing the Wadden Sea's intertidal area. Used
for background layer in the function atl_plot_tag()
}
\keyword{datasets}
