% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{roosts_griend}
\alias{roosts_griend}
\title{Roosts around Griend}
\format{
sf object of the pre-roosts and roosts around Griend
}
\source{
Buffered and smoothed around 60 cm NAP polygon.
}
\usage{
roosts_griend
}
\description{
Roosts around Griend
}
\keyword{datasets}
