% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools4watlas-package.R
\docType{package}
\name{tools4watlas-package}
\alias{tools4watlas}
\alias{tools4watlas-package}
\title{tools4watlas: Process and Visualize High-Throughput Animal Tracking Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for processing and visualising high-throughput animal tracking data. While tailored for WATLAS (Wadden Sea Advanced Tracking and Localization of Animals in real-life Systems), many functions apply to other high-throughput tracking datasets, and study system-specific steps can be adapted easily. In addition, the package enables the integration of environmental data (e.g., water level, bathymetry, or any other raster data), allowing to link movement data easily to other ecological data. Beyond data processing workflows, 'tools4watlas' includes tutorials, and visualization functions that support static plots, interactive plots and animations of movement data, as well as base map data of the Dutch Wadden Sea.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://allertbijleveld.github.io/tools4watlas/}
  \item Report bugs at \url{https://github.com/allertbijleveld/tools4watlas/issues}
}

}
\author{
\strong{Maintainer}: Allert Bijleveld \email{allert.bijleveld@nioz.nl} (\href{https://orcid.org/0000-0002-3159-8944}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Allert Bijleveld \email{allert.bijleveld@nioz.nl} (\href{https://orcid.org/0000-0002-3159-8944}{ORCID}) [copyright holder]
  \item Johannes Krietsch \email{johannes.krietsch@nioz.nl} (\href{https://orcid.org/0000-0002-8080-1734}{ORCID})
}

Other contributors:
\itemize{
  \item Pratik Gupte (\href{https://orcid.org/0000-0001-5294-7819}{ORCID}) [contributor]
  \item Christine Beardsworth (\href{https://orcid.org/0000-0003-1308-1455}{ORCID}) [contributor]
}

}
\keyword{internal}
